# doc/api.yml
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/user/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/track":{"post":{"operationId":"TracksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTrackDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"TracksController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/track/{id}":{"get":{"operationId":"TracksController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"TracksController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTrackDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"TracksController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/artist":{"post":{"operationId":"ArtistsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArtistDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ArtistsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/artist/{id}":{"get":{"operationId":"ArtistsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"ArtistsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArtistDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ArtistsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/album":{"post":{"operationId":"AlbumsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAlbumDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"AlbumsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/album/{id}":{"get":{"operationId":"AlbumsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"AlbumsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAlbumDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AlbumsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/favs":{"get":{"operationId":"FavoritesController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/favs/{type}/{id}":{"post":{"operationId":"FavoritesController_create","parameters":[{"name":"type","required":true,"in":"path","schema":{"enum":["artist","album","track"],"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}},"delete":{"operationId":"FavoritesController_remove","parameters":[{"name":"type","required":true,"in":"path","schema":{"enum":["artist","album","track"],"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}}},"info":{"title":"REST example","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}},"required":["login","password"]},"UpdateUserDto":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}},"required":["oldPassword","newPassword"]},"CreateTrackDto":{"type":"object","properties":{"name":{"type":"string"},"artistId":{"type":"string"},"albumId":{"type":"string"},"duration":{"type":"number"}},"required":["name","artistId","albumId","duration"]},"UpdateTrackDto":{"type":"object","properties":{}},"CreateArtistDto":{"type":"object","properties":{"name":{"type":"string"},"grammy":{"type":"boolean"}},"required":["name","grammy"]},"UpdateArtistDto":{"type":"object","properties":{"name":{"type":"string"},"grammy":{"type":"boolean"}}},"CreateAlbumDto":{"type":"object","properties":{"name":{"type":"string"},"year":{"type":"number"},"artistId":{"type":"string"}},"required":["name","year","artistId"]},"UpdateAlbumDto":{"type":"object","properties":{"name":{"type":"string"},"year":{"type":"number"},"artistId":{"type":"string"}}}}}}